Fix new catalog creation on 'nt'
Code for syncing on 'nt' system would fail in case the catalog
did not exist yet. This code path should be taken only for
existing catalogs, and not when a new catalog is being created.

Problem reported by Ben Brian.

svn path=/trunk/l10n-support/pology/; revision=1419721

+++ pology/catalog.py
@@ -1270,7 +1270,7 @@ class Catalog (Monitored):
         ofl.writelines(enclines)
         if not writefh:
             ofl.close()
-            if os.name == "nt":
+            if os.name == "nt" and os.path.exists(self._filename):
                 # NT does not allow to overwrite on rename.
                 tmpfname2 = self._filename + "~tmpo"
                 os.rename(self._filename, tmpfname2)
